+++ posthog/src/main/java/com/posthog/internal/PostHogPreferences.kt
@@ -52,6 +52,12 @@ public interface PostHogPreferences {
         internal const val ERROR_TRACKING = "errorTracking"
         internal const val CAPTURE_PERFORMANCE = "capturePerformance"
         internal const val PUSH = "push"
+
+        @PostHogInternal
+        public const val PUSH_OPENED_MESSAGE_IDS: String = "pushOpenedMessageIds"
+
+        @PostHogInternal
+        public const val PUSH_SUBSCRIPTION_REJECTED: String = "pushSubscriptionRejected"
         internal const val PERSON_PROPERTIES_FOR_FLAGS = "personPropertiesForFlags"
         internal const val GROUP_PROPERTIES_FOR_FLAGS = "groupPropertiesForFlags"
         public const val SURVEY_SEEN: String = "surveySeen"
@@ -63,6 +69,8 @@ public interface PostHogPreferences {
 
         public val ALL_INTERNAL_KEYS: Set<String> =
             setOf(
+                PUSH_OPENED_MESSAGE_IDS,
+                PUSH_SUBSCRIPTION_REJECTED,
                 GROUPS,
                 ANONYMOUS_ID,
                 DISTINCT_ID,